[08/08/2022] 解决warning C4819
To save a file in Unicode format, in Visual Studio, choose File > Save As. In the Save File As dialog box, select the drop-down on the Save button and choose Save with Encoding. If you save to the same file name, you may need to confirm that you want to replace the file. In the Advanced Save Options dialog, choose an encoding that can represent all the characters in the file—for example, Unicode (UTF-8 with signature) - Codepage 65001—and then choose OK.